# tests/snapshots/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "AdminSessionCookie": {
        "in": "cookie",
        "name": "workflows_admin_session",
        "type": "apiKey"
      },
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "workflows-mcp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/events/v1/sync": {
      "get": {
        "operationId": "sync_events_api_events_v1_sync_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Sync Events",
        "tags": [
          "events-v1"
        ]
      }
    },
    "/api/events/v1/sync/state": {
      "get": {
        "operationId": "sync_events_state_api_events_v1_sync_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Sync Events State",
        "tags": [
          "events-v1"
        ]
      }
    },
    "/api/events/v1/system": {
      "get": {
        "description": "Return minimal SSE stream for v1 events namespace.",
        "operationId": "system_events_api_events_v1_system_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "System Events",
        "tags": [
          "events-v1"
        ]
      }
    },
    "/api/events/v1/watchers": {
      "get": {
        "operationId": "watcher_events_api_events_v1_watchers_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Watcher Events",
        "tags": [
          "events-v1"
        ]
      }
    },
    "/api/events/v1/watchers/state": {
      "get": {
        "operationId": "watcher_events_state_api_events_v1_watchers_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Watcher Events State",
        "tags": [
          "events-v1"
        ]
      }
    },
...
  }
}